# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ArrControl API
  version: 0.1.0
  license: { name: MIT }
servers: [{ url: /api/v1 }]
security: [{ cookieAuth: [] }]
paths:
...
  /events/snapshot:
    get:
      operationId: getLiveEventSnapshot
      description: Returns a durable publication waterline and the REST resources the caller must read before opening the event stream. Reading the snapshot first and the REST projections second prevents reconnect gaps.
      responses:
        '200': { description: Caller-specific snapshot recovery contract, content: { application/json: { schema: { $ref: '#/components/schemas/LiveSnapshot' } } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
  /events:
    get:
      operationId: streamLiveEvents
      description: Streams redacted resource invalidations as SSE. Supply the snapshot cursor as the initial query cursor; browser reconnects use Last-Event-ID. An unknown or expired cursor yields a snapshot-required event and closes.
      parameters:
        - { name: cursor, in: query, required: false, schema: { type: string, maxLength: 64 }, description: Initial cursor returned by getLiveEventSnapshot. }
        - { name: Last-Event-ID, in: header, required: false, schema: { type: string, maxLength: 64 }, description: SSE reconnect cursor; takes precedence over the query cursor. }
      responses:
        '200':
          description: SSE events named changed, cursor, or snapshot-required; heartbeat comments are emitted while idle.
          content: { text/event-stream: { schema: { type: string } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/AccessDenied' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-arrcontrol_session, description: Short-lived opaque access token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
    refreshCookie: { type: apiKey, in: cookie, name: __Host-arrcontrol_refresh, description: Rotating opaque refresh token in a Secure, host-only, Path=/, HttpOnly, SameSite cookie. }
...
  responses:
...
    AuthenticationRequired: { description: 'Authentication is required (`authentication_required`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    AccessDenied: { description: 'The effective authorization does not grant access (`access_denied`).', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
...
    LiveSnapshot:
      type: object
      additionalProperties: false
      required: [version,cursor,generatedAt,resources]
      properties:
        version: { const: 1 }
        cursor: { type: string, maxLength: 64, description: Opaque publication waterline; `origin` when no event has been published. }
        generatedAt: { type: string, format: date-time }
        resources: { type: array, uniqueItems: true, items: { type: string, pattern: '^/api/v1/' }, description: Authorized REST projections to reload before streaming from cursor. }
...
    Problem: { type: object, additionalProperties: false, required: [type,title,status,code,traceId], properties: { type: { type: string, format: uri-reference }, title: { type: string }, status: { type: integer }, detail: { type: string }, traceId: { type: string }, code: { type: string } } }